requirements.txt
# Core ML and Data Science
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.21.0
numpy>=1.21.0
pandas>=1.4.0
scikit-learn>=1.1.0

# Computer Vision
Pillow>=9.0.0
opencv-python>=4.6.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Vector Database
pinecone-client>=2.2.0

# Language Models and AI
openai>=0.27.0
llama-index>=0.8.0
pydantic>=1.9.0

# Web Framework
streamlit>=1.22.0

# Image Processing
ultralytics>=8.0.0

# Utilities
tqdm>=4.64.0
requests>=2.28.0
python-dotenv>=0.19.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0

# Additional Dependencies
omegaconf>=2.2.0 